Hide container when no categories available Fixes #2365
